+++ js/utilities/utilities.js
@@ -7,6 +7,30 @@ export function httpGet(url, callback=console.info, err = console.error) {
   request.send();
 }
 
+export async function httpGetJson(url) {
+   await httpGet(url, function(data) {
+      return JSON.parse(data);
+    return
+   });
+}
+
+export async function fetchFile(file, callback) {    
+  await fetch(file)
+    .then((res) => callback(res))
+    .catch(function (err) {
+      console.error(`fetchFile: Something went wrong with '${file}: ${err}`);
+    });
+}
+
+export async function fetchJsonFile(jsonfile, callback) {    
+  await fetch(jsonfile)
+    .then((res) => res.json())
+    .then((jsonContent) => callback(jsonContent))
+    .catch(function (err) {
+      console.error(`fetchJsonFile: Something went wrong with '${jsonfile}: ${err}`);
+    });
+}
+
 export async function copyToClipboard(stringToCopy, show = null) {
   try {
     await navigator.clipboard.writeText(stringToCopy.toString());